mutt_save_message: rename app -> security_flags
en->email->security is a uint16_t bitmask that contains, amoung other
things, information about application that is used for encryption.
The following assignment operation is misrepresenting the context of
the operation and therefore is misleading

    app = en->email->security

security_flags is a better name for this variable and it has a
proper type SecurityFlags to allow developer navigate to the defintiion
of SecurityFlags type.