mutt_save_message: change delete_original to save_opt
delete_original boolean flag is a misnomer, because it misrepresents the
context of the operation. An email message can be either copied or moved
which is an improtant distinction, while deleting original is just a
detail of the implementation of the move operation.

This patch introduces MessageSaveOpt enum type which is currently a bi-state:
- copy = 0
- move = 1

Values of enum variants are chosen specifically to preserve
compatibility with old boolean flag in case I forgot to update a
call-site somewhere.