Add <view-pager> function to attachment menu
This uses a copiousoutput mailcap entry, or falls back to raw text.

The existing functions do not provide a way to use copiousoutput
mailcap entries exclusively.  The default <view-attach> will use one
for internally supported types if auto_view is configured, but there
is no way to do so for other mime types (such as application/pdf).

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/474d368e859aa86b4adf9e41f1048488ece99c42