Change attach inline forward messages default CH flag
Fix it to be CH_DECODE instead of CH_XMIT.

I believe this was a mistake from a long time ago.  The CH_XMIT flag
is used in neomutt for:
- mutt_save_message() -> set_copy_flags() when decoding
- pre-send conversion to 7 bit
- mutt_make_message_attach() (for index mime-forwarding)
- message bouncing

Using it for *inline* forwarding in the attachment menu doesn't make
sense.  Both include_forward() in send.c, and include_header() in
recvcmd.c use CH_DECODE.  It seems reasonable to decode headers when
inline forwarding from the attachment menu (as is done from the
index), so I'm guessing that was what was meant here too.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/56f4c0b8189594004db37bd31483bb04e19e12db

Co-authored-by: Kevin McCarthy <kevin@8t8.us>