Fix pattern compilation error for ~( !~>(~P) )
The problem is a nested recursive call followed by trailing
whitespace.  Other codepaths ensure trailng whitespaces are processed
when processing the argument, but the parenthesis calls did not.

This meant the outer loop while check:
  while (*ps.dptr)
  {
    SKIPWS (ps.dptr);
    switch (*ps.dptr)
    {
      /* process (~P), for instance */
    }
  }
would return true for the trailing space(s), the SKIPWS would skip
over them, and then switch would try to evaluate '\0'.

I think a better solution would be to rearrange the calls like:
  SKIPWS (ps.dptr);
  while (*ps.dptr)
  {
    switch (*ps.dptr)
    {
    }
    SKIPWS(ps.dptr);
  }
to ensure any trailing whitespace is removed.

However, that's a bit risky for a stable branch commit.  So instead,
just explicitly add the SKIPWS for the ~() thread ops and for the ()
groupings after processing the inside expression.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/564b515cdec7ccc6732f7c299edbcfd7bb028ba6

Co-authored-by: Kevin McCarthy <kevin@8t8.us>