Minor fixes to match_body_patterns()
Change the end condition to compare against buflen.  I prefer this
because we are not directly incrementing offset ourselves, but relying
on the rational return values from regexec.  If the regex library
(incorrectly) returned a nul-match at the '\0' terminator for
instance, the new nul-match logic would assign offset just past the
terminator.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/5cf52cd176101737f2ee60c4ebe5f4603a43e787