Implement backend for create maildir mailboxes
Register a creation function under MxOps as `mbox_create`. This takes a
pointer (to a pointer) to the mailbox to create and returns an enum for
success/failure states which can then be consumed by the browser, e.g.
for MX_CREATE_SYS_ERROR something like `"failed to create mailbox %s:
  %s", name, strerror(errno)`, for MX_CREATE_EXISTS the browser could
prompt the user if they'd like to overwrite etc.

The plan is for this API to be used for creation across other supported
mailbox types.