threads: prevent background sorting from causing refreshes
This is a HACK!  Soon it won't be necessary.

Many functions trigger a (sub)thread to be recalculated.
The thread functions temporarily changed the value of `$sort`, then
restore it afterwards.

Now that the `C_Sort` global variable has been eliminated, it's
necessary to use the Config system to change the value.
Changing the value of `$sort` triggers a notification that causes the
entire Mailbox to be resorted.

This resorting causes unsightly refreshes of the screen.