Allow a -noregex param to folder and mbox hooks
This is helpful if we want to match mailbox that comes from a variable
exactly (or we simply can't be bothered to escape the whole thingie).
As an example, in `folder-hook -noregex $trash set my_foo=bar` we don't
want an expansion of `$trash` such as `imaps://user@gmail.com/[Gmail]/Trash`
to be treated as a regex containing a character class `[Gmail]`.