Do not remove "active" mailboxes
We we are in attach-message mode, we have a stack of two mutt_index_menu
running, and the global Context refers to the mailbox we're browsing to
select the files to attach. The mailbox in the background, where we were
when we started to compose a new message, must be kept alive even if it
is not currently in the Context.

As a side effect, this decontextualizes command_parse.c.