configure: Force to use --pcre2 with libpcre2-posix installed
This bug hit on openSUSE first as it silently pulled in libpcre2-posix
via libncursesw which silently shadows the standard variant of regcomp
and regexec and works a bit differently. Unfortunately this only hits at
runtime with an assertion at the moment.

This patch adds a configure check to make sure that --pcre2 is used
explicitly if libpcre2-posix is found. I tried to make it work by
removing the assertion. Unfortunately the behaviour seems to differ as
well. So I think the safe thing to do is to require --pcre2 if
pcre2-posix is installed.