fix(notmuch): allow windowing with 0 duration
The current implementation of window queries deactivates when
`nm_query_window_duration=0`. This makes windows spanning a single
window timebase impossible. At the smallest allowed duration (1),
windows have a span of [n, n + duration].

For example, `duration=1` and `timebase=month`, generates the window
`date:1month..`, which is "all mail with dates from the previous month
to now". Shifting the window with `vfolder-window-backward` generates
the window `date:2month..1month` which is "all mail with dates from the
month before last and last month".

To generate a window for a single month, the equivalent notmuch search
term is `date:Nmonth..Nmonth` where `N` is the offset from the current
month (0 for current month, 1 for previous month, etc.) This is where the
requirement of `duration=0` comes in. The required window span is
[n, n].

To support zero durations without breaking existing user configurations,
introduce `nm_query_window_enable` to always enable window queries by
disabling the duration <= 0 checks inside windowing code. This is safe
to do as NeoMutt checks that a user has a non-negative integer for
`nm_query_window_duration`. The checks only verify that the value is a
positive integer.

The current behavior of non-zero duration to enable remains intact.
Existing users won't need to make a configuration change so there's no
harm to their user experience.

Fixes #2900

Signed-off-by: Austin Ray <austin@austinray.io>