Try to avoid redundant mx_msg_open calls
This reduces the number of calls to mx_msg_open when viewing a message,
its attachments, or limiting the index, by trying to call mx_msg_open as
soon as possible and pass the resulting Message or FILE pointer down.
This has impact mostly on IMAP w/o `message_cachedir`.

Some net results:
- `<display-message>` then `<view-attachments>` fetches the message once
  instead of twice
- limiting with `|` or `&` with patterns that require a full message
  (e.g., `~M`) fetches the message once instead of once per branch of
  the conditional