Add local_date_header option
Add an option to convert the date used in the Date header into the local
timezone of the sender. This is the current behavior and the option
defaults to true. Unsetting this option causes the date in the Date
header to be formatted using the UTC timezone (aka GMT).

This option is useful for privacy-sensitive users who may not wish to
divulge their sending timezone.

This ports mutt's commit from Gregory Anders
https://gitlab.com/muttmua/mutt/-/commit/4c786d8795fa55eca3685d58edc60361ddf4de37