Directly add full mailbox to GPG search hints
The code previously called (crypt/pgp)_add_string_to_hints(), which
performs a strtok() using " ,.:\"()<>\n" and discards parts of size 3
or less.

This makes sense for personal fields, and random "stuff" entered in
the getkeybystr() prompt.  But it doesn't make a lot of sense for the
mailbox.

There's no advantage I can see in chopping off and discarding the
domain name for getting more relevant results.

Ordinarily, I wouldn't bother fixing this, but in ticket 354 someone
is experimenting with alternative backends, and they would like the
full email address listed in the hints.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/3cf7a1492e8a709adee0e003594d2b58b2953403

Co-authored-by: Kevin McCarthy <kevin@8t8.us>