docs: Document pager_read_delay as a feature
Given that the new config variable gives us feature parity with other
MUAs (for example, Thunderbird has had read delay for some time[1]),
it's worth documenting it as a feature to call attention to this
enhancement.

[1] https://support.mozilla.org/en-US/questions/1246812

Until we have UI events cleanly propagating message status changes
back to the index line, the 'N' flag from the %Z expando of
$pager_index_lines is not reliable.  With pager_read_delay, it would
be nice to have this automated more.

- docs: add a <sect1> feature page