pager: improve skip-headers
Instead of refusing to do anything useful once we have already skipped
to the message body, it is nicer if 'H' (<skip-headers>) would
unconditionally reset us to the start of the email body, no matter
where we have scrolled to.  And since <skip-headers> is a neomutt
feature not in original mutt, we might as well document it as such
(and while at it, fix a typo in the <skip-quoted> function name).