pager: honor mid-message config changes
In mutt, since all config variables are global state, changing them
while in the pager had an effect on later pager commands.  Not so in
neomutt, where our config system read them only once when starting the
pager, and failed to re-read them in the event loop.  Sink the scope
of all affected variables.

For an example, starting in the index, run
: set nopager_stop
<enter>
: set pager_stop
<end>
<space>

Where mutt complained "Bottom of message is shown.", neomutt
accidentally moved on to the next one without this patch.

A bit more subtle: our redraw loop usually manages to do the right
thing when pager_index_lines change.  But OP_DELETE_THREAD and
OP_UNDELETE_THREAD were using the cached value, and could therefore
call menu_queue_redraw() on the wrong portions of the window.