pager: fix N interaction with pager_read_delay
When pager_read_delay is 0, using <mark-as-new> on a message that
started as new will result in that message remaining as marked read
(the message was marked read in act 2, but first is still true).  But
when pager_read_delay is non-zero, whether the message ends up as read
depends on whether N arrived before the timeout, which is confusing.
Better is to remember if we have a pending timeout, and ensure the
message ends up as read as well as cancelling the timeout.

This is a partial fix of https://github.com/neomutt/neomutt/issues/2977