Fix menu.c sigwinch handling order
This fixes problems with a sigwinch that occurs during an external
process, such as during a query menu lookup, or after exiting the
background-edit landing page.  Without this change, the user would see
a misdrawn screen and would need to press a key first.

Change it so that the SigWinch flag is handled before reading input,
so that the page is automatically redrawn.  This matches the behavior
in the index, pager, and background-edit process list menu.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/f5ccd687348a8a97bb442d031bdd4e68091615c0

Co-authored-by: Kevin McCarthy <kevin@8t8.us>