threads: Expose $use_threads as %T in $status_format
Although it is generally obvious from glancing at the index whether
threading is in use, we regressed in the portion of the status line
that changed between "(date/date)" and "(threads/date)" when using
'ot' once $use_threads has been set.  The solution to this is to add a
new expando, %T, and give it the smarts to resolve to the actual
threading method ("flat", "threads", "reverse", and never "unset")
regardless of whether we are threading via legacy or modern config.

It also makes sense to make this expando conditional, since it is
highly desirable to see "(threads/last-date/date)" when copying
Thunderbird sorting (where top-level threads sort differently than
subthreads within a thread), but omitting "flat/" makes sense when
threading is off.  This also entails an update to our default for
$status_format, with accompanying documentation.