sort: fix regression on config observers
Fixes: https://github.com/neomutt/neomutt/issues/2988

Commit faf75108 accidentally lost the ability to trigger the three
global re-sort flags (R_RESORT and friends) on associated
configuration changes.  This is visible in things like changing
$sort_aux and $strict_threads no longer triggering an immediate
recomputation of the thread tree.  As the decision on whether to
re-sort a threaded index is not strictly a menu feature (although
after the re-sort, the menu will indeed be redrawn), it seems better
to move that code into the index config observer, especially since the
index already owns config_reply_regex() which also sets OptResortInit
(which is slightly odd since $reply_regex is tagged with R_RESORT but
not R_RESORT_INIT, but that's a story for another day).

Note that faf75108 also dropped all handling for R_INDEX and R_PAGER
handling (which arguably would live better in index/index.c and
pager/pager.c's configure_observer, respectively, rather than in the
generic menu/observer.c), but my attempts to demonstrate a regression
using those bits (for example, changing $to_chars) did not spot any
loss in behavior (the index was still immediately redrawn with the new
characters).  We may still find cases where we need to fix various
observers to honor more R_ flags, and/or get rid of the R_ flag scheme
altogether in favor of the observer pattern listing explicit config
variable names, but that can be later cleanup; this patch is focusing
just on the observable regression in re-sorting.