muttlib: add mutt_qsort_r
Being able to sort data with reference to auxiliary information passed
in through a local parameter rather than requiring non-local context
through a global variable makes code easier to read.  qsort_r() will
be added in a future version of POSIX [1] with the same signature
present in glibc, and nicely fills this role, but is not portable yet,
so we have to write a shim instead.  Various FreeBSD releases ship
with an incompatible signature for qsort_r (the opaque variable passed
in a different parameter); FreeBSD 12 is hopeless (we'd have to shim
that reversed argument), but starting with FreeBSD 13, we can use
qsort_s instead (however, note that C11 mentioned qsort_s as optional,
and Windows has qsort_s with yet another incompatible signature).  And
for platforms that lack either re-entrant form, as long as our signal
handlers and comparators don't themselves try to call a secondary
sort, we don't care about re-entrancy, and can isolate the use of
external data to static variables in one file rather than global
variables across multiple files.

(Writing a qsort_r replacement on top of qsort where re-entrancy does
matter would require a mutex and shenanigans to support a stack of
nested calls, which gets a lot more complex.)

The unit test will hopefully catch any other special-casing we need to
perform in writing our wrapper for other systems with non-standard
signatures of either qsort_s or qsort_r.