sort: Determine thread sort_key more directly
Reusing the qsort_r callback compare_threads outside of qsort_r is
convoluted compared to just inlining the direct call to
mutt_compare_emails.  Making this change also allows us to add an
assert that compare_threads is now only used only between siblings,
which will ease the efforts in the next patch to sort subthreads
differently than overall threads.

To see that the conversion is sane, note that logically, pre-patch was
checking either:
c_sort & SORT_REVERSE: 1 * compare(a, b, SORT_REVERSE) > 0
!(c_sort & SORT_REVERSE): -1 * compare(a, b, !SORT_REVERSE) > 0
Always passing SORT_REVERSE to the comparison gets rid of the need to
multiply by the result of a ternary.