Add a hard redraw after oauth authentication
The sample contrib program can use gpg to decrypt a token.  This may
pop up a curses agent prompt, requiring a hard redraw.

It may be that we should be calling a hard redraw after any fork/exec.
I don't know if this will cause more problems, so for now I'll just
patch this one place.  But that is something to consider.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/b53593393bab8bc0733a39a772219cf92f5a8b49

Co-authored-by: Kevin McCarthy <kevin@8t8.us>