notmuch: implement query window reset function
Implement functionality for resetting the notmuch query window back to
the present. If a user moves the window backwards n times, they must
move the window forward n times or close and reopen the mailbox to
return to present. Having a reset functionality improves the quality of
life for the user.

To do this, introduce a new user function, `vfolder-window-reset` and
corresponding opcode. When a user executes `vfolder-window-reset`, it
set the position, stored in the `nm_query_window_current_position`
config. value, to 0. The rest of the logic follows the same flow as
`vfolder-window-forward` and `vfolder-window-back`.

Feature request originated in issue #2972 as an additional request. This
commit does not fully close that issue.