index: simplify notmuch query window operations
All three query window operations share the same control flow except for
the notmuch backend call. To reduce the amount of duplicate code, merge
the operations into a single code block with a switch for the backend
call. This allows the operations to share the same guard clauses and
folder change logic while preventing any errors when a common change
made in one operation isn't propagated to the other two.