pager: kill menu
Despite being built on a Menu, the Pager doesn't really use the Menu.

The Pager implements all the 'motion' functions that are handled by the
Menu on other screens.

The Pager used to save flags to `Menu->redraw`.  Now they've been moved
to `PagerPrivateData->redraw`.

The only place where NeoMutt is expecting the Pager to have a Menu is in
`menu_get_current_type()`.  This temporary workaround will be eliminated
soon.