threads: Refactor code that cares about thread style
An upcoming patch will introduce the $use_threads config option to
make the choice of threading orthogonal from the rest of $sort
decisions.  In preparation for that, all code that currently reads
c_sort directly for the purpose of making decisions based on threading
now instead goes through a new helper function, so that it becomes
easier to eventually resolve the priority between the config
variables.

Remaining sites that still use cs_subset_sort(NeoMutt->sub, "sort")
are concerned with other types of sorting, not threading.