config: Kill pager_validator
pager_validator() was added in commit cbc52d9f (in release 20191025)
to mimic what upstream mutt does with its CHECK_PAGER macro, to
prevent ':set sort=date' and friends from the pager.  But as written,
our validator caused a regression: it ALSO prevents the operation of
<sort-mailbox>, which is something that upstream mutt permits.

Perhaps there was an ancient technical restriction where re-sorting
while in the pager could cause the current message number to change
due to its new order, and which may have been visible via any pager
operation that jumps into the index then back into the pager not
landing in the same place. But it appears that such a restriction has
long been lifted in the past, with only the prohibition against
setting certain variables left behind only out of inertia.

I see no continuing technical reason to prevent re-sorting the mailbox
from the pager; whether or not pager_index_lines is non-zero, we still
end up displaying the correct message, and when $sort is set, the index
view properly updates to reflect the changed sort.  Thus it is time to
kill this misfeature inherited from mutt, and allow the user to change
sorting variables even from the pager.

In testing this, we do have at least one problem that still needs
fixing: ':toggle strict_threads' does NOT update the index half of the
pager view if there is at least one message that forms a pseudo-thread
by subject line but not by reference headers; it is not until the
subsequent <next-undeleted> or <quit> back to the index view that the
index is then shown correctly.  Probably a hole in
index_config_observer() not properly sending a redraw event on
R_RESORT_INIT config changes.