threads: Document $use_threads default
It turns out that I misspelled the default in commit 3bc1f623, but no
one noticed because docs/config.c isn't compiled.  What's more, we
don't handle defaults for DT_ENUM very well in makedoc.c; it requires
manual updates for every default value anyone might use (we might
improve that in the future).  With this patched, the manual now
documents 'unset' as the default.