Enable nonl() mode for ncurses
This patch is based on the patch from Mike Frysinger.

This version creates a new key symbol, <KeypadEnter> for the KEY_ENTER
value returned by getch().  According to Vincent Lefèvre's research,
that corresponds to the "key pad" Enter key.  If KEY_ENTER is not
defined, it binds the symbol to "\n".  It does this after the
KeyNames[] definitions of <Enter> and <Return> so they remain the
default shown by <what-key>.

Since Mutt already accepts KEY_ENTER at the y/n prompt, add it for all
the built-in operations too.

Bindings with "\r" and "\n" in the binding_t structures don't need
"<enter>" and "<return>" to be re-specified in km_init(), so remove
those.  Instead add the "<KeypadEnter>" bindings there.

Remove unnecessary Alias menu bindings for OP_GENERIC_SELECT_ENTRY in
km_init().

As a note: I don't believe non-extended key symbol bindings need to be
bound inside km_init().  I'm not clear why this was done, or what
purpose it serves.  I'll do a bit more research, but if I can't
uncover any issues, will move them inside the binding_t structures in
another commit.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/f4ff768c2764502a974042e1fe9d861a7a92fbcf

Co-authored-by: Kevin McCarthy <kevin@8t8.us>