Break the menu OPS/funcs and keybindings into separate tables
Move the "extra" keybindings added manually inside km_init() into the
keybinding tables instead.

This allows the documentation in the manual to be generated correctly.

It also corrects a small problem with duplicate OPS in the struct
binding_t table.  The exec completion and unbound list in the help
menu iterated the active binding_t table and assumed each OP appears
only once.  When '\n' and '\r' were both added to the binding_t table,
this broke the assumption and caused those to misbehave.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/4021ab59f021abd84742ae36895158c7129cbb38

Co-authored-by: Kevin McCarthy <kevin@8t8.us>