Add $compose_confirm_detach_first, default set
With the default keymap, hitting 'D' in the compose menu will remove
parts and unlink owned files.  Mutt prevents removing all the parts,
but if you have added attachments, this makes it very easy mis-type
and accidentally nuke the message you've been composing.

This option only checks for the first part - which is highly likely
the main composed message.  It might be possible to try and make the
check more intelligent, but I think a simple check will go a long way
towards preventing disaster.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/7358075a415211ea1ea83b75adc4dbdfcfde85ff