Don't use subtraction for qsort numeric value comparisons
Subtraction can overflow, resulting in incorrect sorts.  This is
especially a concern for the date and size sorting, whose fields are
greater than an 'int' size.

The index values should be okay, but it's better to be consistent and
avoid any possible issues.

Define a macro, mutt_numeric_cmp(), that uses direct comparison rather
than subtraction.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/d0d40a33f1400269251d06c56ef331482c6b99be

Co-authored-by: Kevin McCarthy <kevin@8t8.us>