fix(index): find next mailbox with unread mail
Existing implementation of `next-unread-mailbox` finds the next mailbox
with new mail. This contradicts the function's name and
documentation[0], which states:

   The (by default unbound) function '<next-unread-mailbox>' in the
   index can be used to immediately open the next folder with unread
   mail (if any).

To bring implementation in line with the name and documentation,
introduce 'mutt_mailbox_next_unread()', which mirrors
'mutt_mailbox_next()' except it for unread check mail and omission of
stat checking, and call it.

[0] https://neomutt.org/guide/advancedusage#13-2-%C2%A0polling-for-new-mail