Improve pattern compiler whitespace skipping
Commit 564b515c fixed a problem with nested patterns, where trailing
whitespace could cause a compilation error.  That fix added missing
SKIPWS() calls in a couple branches.

However, the real problem is that the loop doesn't ensure whitespace
is skipped *before* checking for *ps.dptr.

Rather than relying on all the case branches (and "eater" functions)
to skip the whitespace after processing, rearrange the logic so the
SKIPWS() is after the switch statement.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/534f3ae044ff05a6a59ed25f184098a3305b1226