fix(notmuch): null DB pointer on open failure
This commit nulls 'db' if there's an error in opening the database. The
discussion on #3082 revealed that a failure to open the database is not
guaranteed to NULL the database pointer. So our existing NULL checks in
'nm_db_do_open()' won't function and we may return an invalid non-NULL
database pointer.