fix(notmuch): check database dir for xapian dir
This commit modifies 'nm_db_get_mtime()' to also check for a 'xapian'
directory in the provided database directory. Starting in notmuch 0.32,
databases not located in the mail root might be stored outside a
'.notmuch' sub-directory so 'nm_db_get_mtime()' will fail and cause
'nm_mbox_check()' to fail. As user's may prefer the separate database
and mail directories, we should accommodate that.