Fix mutt_ts_capability() fallback list loop
The for loop termination check should be '*termp', not 'termp'.  termp
starts as &known[0], so to check for the NULL terminator, we need to
look at *termp.

Since *termp is now checked in each loop, it no longer needs to be in
the if check inside the loop, so remove it there.

This hasn't been blowing up because the mutt_strncasecmp() check
inside the loop was also incorrect.  It should be scanning for a
matching entry, but was looking for non-zero (nonmatching) retval.
Since term can't match both the first and second entries, it would
always return 1 and would never get to the loop NULL terminator.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/f58795c9ddf354eb7bb0412decb894fae3a19571