Fix $reply_regexp default value
The default value happened to work, but didn't make logical sense.  It
was using a bracket expression that *included* a backslash, and
left-bracket, 0-9, and again a backslash.  The first ']' closed the
bracket expression, and then '+' matched that 1 or more times,
followed by ']'.  Fortunately the whole parenthisized expression had a
'*' so it would repeat matching, but as far as I can tell, it was
working accidentally.

Because even the fixed version is hard to understand, add some
explanation of the parts to the documentation.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/cd3a5c8510303b4efb9e6ef882a6a179ae92e92c