newsrc: bugfix; nntp_user and nntp_pass ignored
When connecting to an nntp server such as news.eternal-september.org,
there are three ways to connect: First, without a username/password pair,
in which case, it displays a few posts on the server with information
about how to connect correctly. Second, with the username/password
encoded in the url, which works perfectly after #3191. And lastly, by
specifying a username/password pair in nntp_user and nntp_pass, in which
case, the variables are completely ignored in the connection to the
server. The behavior is puzzling, and this patch proposes a fix.

Signed-off-by: Ramkumar Ramachandra <r@artagnon.com>