lens: introduce a new Mailbox type
neomutt faces one stark defeciency compared to modern web-based email:
in an email thread for a received email, the emails that are sent by the
user are missing. In order to fix this, we propose a new Mailbox type
dubbed "lens", that merely acts as a lens into multiple existing
Mailboxes.

This patch is in its very early stages, and is more of a mockup than any
functioning code.