fix crash when forwarding attachments
To forward an attachment, NeoMutt makes a copy of the Body.
Unfortunately, mutt_body_copy() uses memcpy() which duplicates pointers too.
When the email is sent, the Body is frees, leaving dangling pointers in
the original Body.

Ensure that all heap-allocated objects are duplicated, or their
pointers are set to NULL.