Clarify description of $local_date_header
When I reexamined the documentation of the $local_date_header option, I
wasn't very confident about what it actually does. Much of what I did
get from it came from a vague recollection of the discussion around when
it was added, and I referred back to the commit which added this to
solidify that understanding.

Viewing this as a user (without looking at the internal implementation),
I don't think the date is being converted, to me that implies that it
was received from somewhere else in some other format; rather I'd view
it as being initially generated in the local timezone.

The reference to the "sender's timezone" makes me think that this would
be operating on dates in messages received from other people; I've
changed that to "your" to help clarify that it's about messages created
locally.

Finally, add information about what is done if this option is unset,
along with some reasoning on why someone might want to unset this.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/7d2a53ee77bd336d960d1ba91e15681d9bc4f346

Co-authored-by: Aaron Schrab <aaron@schrab.com>