Fix mutt_paddstr() to properly filter unprintable chars
The original version of this function had no "replacement character"
functionality, so it simply directly called addnstr() to display the
characters if everything was okay.

Commit a080fd35 added replacement logic, similar to
mutt_format_string(), but forgot to change addnstr to use the
replacement character.

This means garbage characters could goof up the mutt display, for
things such as the subject in the compose menu.

We could add calls to wcrtomb() like in mutt_format_string(), but
mutt_addwch() already does this, and properly calls wcrtomb() a second
time to add a shift sequence if needed.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/70958893f85d1ca78a7b07a446cfa78779848d1a