Make free'ing a mailbox easier for the caller
Let's just try and free a mailbox whenever we're done with it, and
leave it to mailbox_free to decide when it's actually safe to do.

In this PoC, safe means the mailbox isn't open and isn't visible. This
can probably cause a few leaks, e.g., when quitting NeoMutt.

Force mailbox deletion on removal from account

Create hidden mailboxes, make them visible on "mailboxes"