Force SMTP authentication if smtp_user is set
Chicken and egg: MUTT_ACCT_USER gets set the first time the username is
requested, but SMTP requests the username (i.e., does authentication)
only if MUTT_ACCT_USER is set.

This was causing an smtp_url without the username part to never trigger
an authentication even when smtp_user was set.