Do not insert an actual key mapping for noop
`unmacro * dd` adds a binding from `dd` to `OP_NULL` (i.e., `noop`) to
all menus, including the editor. This `dd` mapping will keep matching
text typed into the editor, so if we type `abcde` the `d` will be
matched and eaten, resulting in `abce`.

This is a quick fix that solves the `unmacro *` case of #3288, but the
problem is more profound. If we want the editor to jump to the
begin-of-line on `bl`, we might try to `macro editor bl <bol>`. This
works, but this also eats any `b` we type in the editor. The editor is
special in this regard, because unlike other dialogs we want typed keys
to actually appear. I don't have a solution for this yet.

Fixes #3288