create resolve_email()
Create a central function to handle all the $resolve behaviour.

First, $resolve is checked.  If it's false, nothing is done.

If action is required, the new position will be calculated using the ResolveMethod.
On success, the Menu will be updated which will send a notification.
On failure, a notification (NT_INDEX_EMAIL,NULL) will be sent.