Fix forwarding text attachments when honor_disposition is set
`mutt_can_decode` is used in `inline_forward_attachments` to decide
which parts of a message need to be attached when forwarding an email.
If we don't consider skipped types such as signatures, this set must be
equal to the set of parts that wasn't included in the email text.
However, using `mutt_can_decode` is not enough, because parts with
disposition = attachments that could be decoded might have been skipped
during the formation of the email text because of `honor_disposition`.
Hence, we need a common piece of code to decide if a part is handled
inline or in an attachment.

An alternative to this would be to mark the parts that were handled in
the main body and include all others. I am not sure what is the best
approach.

Also, please double check my logic here.