Fix uudecode buffer overflow
mutt_decode_uuencoded() used each line's initial "length character"
without any validation.  It would happily read past the end of the
input line, and with a suitable value even past the length of the
input buffer.

As I noted in ticket 404, there are several other changes that could
be added to make the parser more robust.  However, to avoid
accidentally introducing another bug or regression, I'm restricting
this patch to simply addressing the overflow.

Thanks to Tavis Ormandy for reporting the issue, along with a sample
message demonstrating the problem.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/e5ed080c00e59701ca62ef9b2a6d2612ebf765a5

Co-authored-by: Kevin McCarthy <kevin@8t8.us>