Fix off-by-one error in progress during IMAP fetch
We have three nested loops here:

1. separate chunks (see imap_fetch_chunk_size)
2. messages within one chunk + handle the datastructures for loop 3
3. responses from the server until we get a message

Loop 2 was entered a last time after all messages were fetched. I have
moved the call to progress_update to where we are sure we are handling a
new message, moved the exit condition at the end of loop 3 and turned
loop 2 into a while (true) to make it clear that the exit condition is
elsewhere.

Fixes #3389