Fix mbrtowc() error handling in mutt_which_case()
The function did not reset the increment value on a -2 return value.

Increase the maximum conversion size to the string length, and check
for -2.  Since we're looking at the whole string, we can then just
terminate the loop on either value, assuming a case-sensitive search.

mbrtowc() will return -2 if passed n==0, so add an explicit check for
the end of string and a positive n count.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/f8336984f055b151f0b212b32190249ac6f770e6