Implement 'prompt_user' command
'prompt_user' is a new command that takes a variable and, optionally, a
message implemented by parse_prompt. It prompts the user if a message
was given, and stores their response in the variable. The user can abort
giving an input, which leaves the variable unchanged.

This command takes a number of flags:

* -hidden: hides the user input whilst it is being typed
* -file: enables file/directory name completion on the user input
* -anykey: stop taking user input after the first keypress

Flags can be anywhere in the command, and the target variable is the the
first non-flag token. If a second non-flag token is given, it is used as
the user prompt. The flags -hidden and -file conflict, and the one given
last will be used instead of the other.