Fix scrolling when handling SIGWINCH in the index
The index had some (very very old) code which reset the scroll top
during SIGWINCH handling.  This caused a recomputation of the top,
which caused the index to jumble around randomly during resizing.

The last few commits added SigWinch flag setting whenever
mutt_endwin() was called, which meant this jumbling would occur much
more often.  (For example when piping a message.)

After looking more closely, this top reset seems to be unnecessary.  A
full redraw needs to properly handle the case where "current" is
outside the visible range, because a resize could occur in a called
menu.  Additionally, the menu.c code does just fine without this, and
is used for all the other standard menus in Mutt.

Thanks to Vincent Lefèvre for helping test the SigWinch changes, and
reporting this regression.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/9cfa36ea4b388d91a8b9af181866ac8412ea4666