refactor: Remove version check in `notmuch/db.c`
Having `msg` be a null pointer has negligible runtime penalty in the
rare case of a user linking against an old version of notmuch. It's
worth the trade-off to improve readability of this function by removing
the version checks. In addition, the only usage of `msg` post
initialization has a null check.