Fix gpgme crash when listing keys in a public key block
The gpgme code handling classic application/pgp assumed each key would
have a uid.  Change it to check for a missing uid list.

Also change it to list every uid (instead of only the first), and to
put each one on a "uid" line in the output.

The output is only for display, so the format change won't affect
other parts of the code.

Thanks to Mikko Lehto for the high quality bug report, detailing the
exact place of the crash with a reproducing example and a workaround
patch.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/48b6ea32e21db8b580cd3ca8c346c3e2c22756f6