Do not return a single "." for a domain name
When we compute the hostname, we get the canonical name from addrinfo, take the part after the first ".", and use that as a domain name.
It might happen (see #3568) that addrinfo.ai_canonname ends in a ".".
In this case, we don't want to end up constructing "<hostname>.", so let's skip an empty domain name altogether.

Fixes #3568