mutt_oauth2: Print access token request message
There are cases when using the 'authcode' grant where the authorization
request will succeed but the access token request will fail (for
example: if the user's web browser and terminal use different proxy
settings).  The current implementation of the script does not inform
the user that the authorization token is being exchanged for an access
code, with the result that it can appear that a request has both
succeeded (according to the browser) and failed (according to the
terminal output) simultaneously.  Add a message to inform the user that
a second request is being made so there is less potential for confusion.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/7c4fa47888d0dee4529f3a521bec76420528776e