Clarify reference to NeoMutt guide in $header_cache
The documentation of $header_cache reference a section "caching" which
only exist in the NeoMutt Guide but not in the man page.  Clarify that
this is reference to the guide.

Note: We did not change the name ("caching") of the reference, even if
the referenced section is called "Local Caching" in the chapter
"Optional Features".  The reason to keep this wrong name is to get
a proper linking as "caching" is the HTML anchor name of the target
section.