attachments: Document the special treatment of the 1st MIME-part
The first MIME-part is treated slightly different when it comes to
detecting it as an attachment.  This is because almost always is the
first MIME-part the message text and hence not an attachment.

Document these exception in the manual and the neomuttrc in the section
where the attachment command is discussed.