Document handling of `source` command inside hooks
Since eb48fd1368b5d821a74708963aff45f61b709a4b hook store the file
context they are defined in and restore that context when they execute
their commands.  As a result a source command inside a hook is executed
with the context of the file the hook was defined in.  Thus, relative
filenames are interpreted relative to the source file the hook was
defined in.