Move pager_read_delay check to after the blocking km_dokey()
The blocking part of the event loop of the pager is waiting for inputs
by the user.  When we check for an expired timer, we should do this
after this waiting.

Since handling the key event might alter the internal state such as the
currently selected message, we must handle the expired timer events
before the key events otherwise we might ending up operating on the
wrong message.