Pager: Remove post event-loop check for pager_read_delay
The event loop of the pager handles all sorts of events among them the
events of the index which move to a different message,
e.g. <next-entry>.

If the pager_read_delay times out after we have moved to a different
message in the pager, that message is marked as read.  By removing the
read delay check after the loop we prevent the message moved to from
being marked as read.