Postpone messages with Mutt-* headers instead of X-Mutt-*
When postponing a message use the new Mutt-* variants of our custom
headers, e.g. Mutt-Fcc instead of X-Mutt-Fcc.

Mutt can already recall messages with Mutt-* headers, so this is now the
writing side of the transition from X-Mutt-* to Mutt-*.

The reasons to prefer "Mutt-Fcc" over "X-Mutt-Fcc" are detailed in
[RFC 6648] "Deprecating the "X-" Prefix and Similar Constructs in
Application Protocols".

[RFC 6648] https://datatracker.ietf.org/doc/rfc6648/