Support Mutt-Mix and X-Mutt-Mix pseudo header when composing
Allow user to specify a Mutt-Mix and X-Mutt-Mix pseudo header when
composing which sets the mixmaster chain.

The X-Mutt-* style headers were used by postponed message and are
supported in case the user is familiar with this.

The new Mutt-Mix header is preferred over the oldish X-Mutt-Mix ones for
the reasons outlined in [RFC 6648].  Hence, we only document the
Mutt-Mix variant in the manual and omit the X-Mutt-Mix version.

[RFC 6648] https://datatracker.ietf.org/doc/rfc6648/