Support Mutt-* pseudo headers in postponed mails
When recalling a postponed message NeoMutt recognises some special
pseudo headers starting with "X-Mutt-", namely:
"X-Mutt-References", "X-Mutt-Fcc", "X-Mutt-PGP", "X-Mutt-SMIME", and
"X-Mutt-Mix".

Change the implementation to also accept the headers without the "X-"
prefix.  The reasons for this are detailed in [RFC 6648] "Deprecating
the "X-" Prefix and Similar Constructs in Application Protocols".

[RFC 6648] https://datatracker.ietf.org/doc/rfc6648/