Support for Mutt-* and X-Mutt-* pseudo headers when composing
When composing messages NeoMutt recognises the special pseudo headers
"Fcc:", "Attach:", "Pgp:", and "Smime".  Expand this list to also
recognise the headers with "Mutt-" and "X-Mutt-" prefixed:
"Mutt-Fcc", "Mutt-Attach", "Mutt-Pgp", and "Mutt-SMIME", as well as
"X-Mutt-Fcc", "X-Mutt-Attach", "X-Mutt-Pgp", and "X-Mutt-SMIME".
(all matches are case insensitive.)

The later versions are unambiguous and thus should not collide with any
other custom header from other applications.

The X-Mutt-* headers were used by postponed message and are supported in
case the user is familiar with this.

The new headers Mutt-* are preferred over the oldish X-Mutt-* ones for
the reasons outlined in [RFC 6648].  Hence, we only document the Mutt-*
variants in the manual and omit the X-Mutt-* versions.

[RFC 6648] https://datatracker.ietf.org/doc/rfc6648/