Remove usage of MUTT_DATE_NOW
Replace the constant MUTT_DATE_NOW with the function mutt_date_now().
The former was an arbitrary date which was checked for by
mutt_date_localtime() and mutt_date_gmtime() to be replaced by the
current timestamp.  Better pass directly pass the timestamp
corresponding to now to those two functions.

This allows NeoMutt to handle the date Wed 31 Dec 22:13:21 1969, which
was MUTT_DATE_NOW, correctly.