Generate the message id using purely random data
Only use random bits to generate a message id.  This slightly enhances
metadata leaks.

The Length of the message id is chosen such that it is maximal and fits
in the recommended 78 character line length for the headers Message-ID:,
References:, and In-Reply-To:, this leads to 62 available characters
(excluding `@` and `>`). Since we choose from 32 letters, we have
N = 32^62 = 2^310 different message ids.

Example:

   Message-ID: <12345678901111111111222222222233333333334444444444@123456789011>
   References: <12345678901111111111222222222233333333334444444444@123456789011>
   In-Reply-To: <12345678901111111111222222222233333333334444444444@123456789011>

The distribution of the characters to left-of-@ and right-of-@ was
arbitrary.  The choice was made to put more into the left-id and shorten
the right-id to slightly mimic a common length domain name.

Technically this violates the requirement of the message-id being
globally unique [RFC 5322, section 3.6.4].  However, this cannot be
ensured either way, e.g. different MUA on the same machine might
generate the same Message-ID, with this approach, the odds for
a collision are negligible:  The number of message id to generate to
have a collision probability of p is about

   n(p) ~= sqrt(2*N*ln(1/1-p))

with N = 2^310 we get

   n(0.5)   ~= 2^155 > 10^46
   n(0.001) ~= 2^150 > 10^45

[RFC 5322] https://datatracker.ietf.org/doc/rfc5322/