debug: fix email debugging
Fix possible null pointer dereference in dump_envelope()

Test if the passed envelope is a null pointer, if so print "NULL
pointer" otherwise dump the envelope.

Let dump_*() functions print if they received a null pointer

These methods are used for debugging, it is therefore sensible to have
an indication a) weather that dump_*() statement was run and whether it
was passed a null pointer.

Co-authored-by: Richard Russon <rich@flatcap.org>