Decrypt an S/MIME encrypted, postponed message when recalling
An message which was marked for S/MIME encryption and was then
postponed.  Is encrypted with S/MIME if the config variable
$postpone_encrypt is set.

When recalling such a message NeoMutt did not decrypt that message.  By
coincidence it was later decrypted by the MIME handling mechanism.
However, this left the message corrupted, e.g. wrong charset encoding,
wrong filename.  What happened was, that the decrypted content was
combined with the encrypted headers + some modification to the headers.

Fix this by decrypting the message when recalling a postponed massage.
This is the same as we already do for PGP encrypted messages but somehow
S/MIME was forgotten here.