xdg: Add functions for dealing with the XDG environment variables
Add function to query the environment variables defined in the XDG Base
Directory Specification. The functions return the environment variable
or if it is not set, falls back to the default defined in the
specification.

Note that we do not define XDG_RUNTIME_DIR as this has no named default
path. To avoid special cases for this variable, we simply never define
it (the current code does not use it).

A version 0.8 of the specification also describes a directory for
executables but gives it no official name. We added it under the name
XDG_BIN_HOME but commented the related constants.