Add config variables representing the XDG environment variables
Add config variables which reflect the current directory values for the
XDG Base Directory Specification.

The configuration supports accessing environment variables with the
usual `$` syntax. However, XDG environment variables are special in the
sense that their absent means a particular default value, e.g. if
XDG_CONFIG_HOME is not set, then ~/.config should be used instead. This
makes it cumbersome to deal with.

Add the config variables

* $xdg_app_data_home
* $xdg_app_config_home
* $xdg_app_state_home
* $xdg_app_cache_home

which is the value given by the environment variables or the default
value. Their value is the neomutt specific directory not the directory
given by XDG_abc_HOME, e.g. $xdg_app_config_home is ~/.config/neomutt.